separate key/val interface to a different module
