+++ lua-visual-debug.lua
@@ -65,51 +65,6 @@ local KERN = node.id("kern")
 local PENALTY = node.id("penalty")
 local GLYPH = node.id("glyph")
 
-local keyval = require('luakeyval')
-local process_keys = keyval.process
-local scan_bool = keyval.bool
-local scan_string = token.scan_string
-local scan_float = token.scan_float
-local inner_keys = {
-    hlist = {
-        show = {scanner = scan_bool},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float}
-    },
-    vlist = {
-        show = {scanner = scan_bool},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float}
-    },
-    rule = {
-        show = {scanner = scan_bool},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float}
-    },
-    disc = {
-        show = {scanner = scan_bool},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float}
-    },
-    glue = {
-        show = {scanner = scan_bool},
-    },
-    kern = {
-        show = {scanner = scan_bool},
-        negative_color = {scanner = scan_string},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float}
-    },
-    penalty = {
-        show = {scanner = scan_bool},
-    },
-    glyph = {
-        show = {scanner = scan_bool},
-        color = {scanner = scan_string},
-        width = {scanner = scan_float},
-        baseline = {scanner = scan_bool}
-    },
-}
 local params = {
     hlist = {show = true, color = "0.5 G", width = 0.1},
     vlist = {show = true, color = "0.1 G", width = 0.1},
@@ -120,43 +75,6 @@ local params = {
     penalty = {show = true},
     glyph = {show = false, color = "1 0 0 RG", width = 0.05, baseline = true},
 }
-
-local function set_params(key)
-    local vals = process_keys(inner_keys[key])
-    for k,v in pairs(vals) do
-        params[key][k] = v
-    end
-end
-
-local function onlyglyphs()
-    for _,v in pairs(params) do
-        v.show = false
-    end
-    params.glyph.show = true
-end
-
-local outer_keys = {
-    hlist = {scanner = function() return true end, func = set_params},
-    vlist = {scanner = function() return true end, func = set_params},
-    rule = {scanner = function() return true end, func = set_params},
-    disc = {scanner = function() return true end, func = set_params},
-    glue = {scanner = function() return true end, func = set_params},
-    kern = {scanner = function() return true end, func = set_params},
-    penalty = {scanner = function() return true end, func = set_params},
-    glyph = {scanner = function() return true end, func = set_params},
-    onlyglyphs = {default = true, func = onlyglyphs}
-}
-
-do
-  if token.is_defined('lvdset') then
-      texio.write_nl('log', "lua-visual-debug: redefining \\lvdset")
-  end
-  local function_table = lua.get_functions_table()
-  local luafnalloc = luatexbase and luatexbase.new_luafunction 
-    and luatexbase.new_luafunction('lvdset') or #function_table + 1
-  token.set_lua('lvdset', luafnalloc)
-  function_table[luafnalloc] = function() return process_keys(outer_keys) end
-end
 
 local function math_round(num, idp)
   if idp and idp>0 then
@@ -312,5 +230,6 @@ end
 
 
 return {
-  show_page_elements = show_page_elements
+  show_page_elements = show_page_elements,
+  params = params
 }
